+++ changes.md
@@ -14,4 +14,5 @@
   - add `module Web.Audio` and `module Web.Audio.Parameter`
   - add `module Web.Socket`
   - add `module Web.LocalStorage`
+  - add `module Web.GeoLocation`
   - TODO example audio speed does not seem to work. Make sure it does!